requirements.txt
Package                                  Version
---------------------------------------- -----------
annotated-doc                            0.0.4
annotated-types                          0.7.0
anyio                                    4.12.1
asttokens                                3.0.1
attrs                                    25.4.0
Authlib                                  1.6.6
backoff                                  2.2.1
beartype                                 0.22.9
cachetools                               6.2.5
certifi                                  2026.1.4
cffi                                     2.0.0
charset-normalizer                       3.4.4
click                                    8.3.1
cloudpickle                              3.1.2
colorama                                 0.4.6
comm                                     0.2.3
cryptography                             46.0.3
cyclopts                                 4.5.1
debugpy                                  1.8.20
decorator                                5.2.1
diskcache                                5.6.3
distro                                   1.9.0
dnspython                                2.8.0
docstring_parser                         0.17.0
docutils                                 0.22.4
email-validator                          2.3.0
exceptiongroup                           1.3.1
executing                                2.2.1
fakeredis                                2.33.0
fastapi                                  0.135.1
fastmcp                                  2.14.4
filelock                                 3.24.3
fsspec                                   2026.2.0
googleapis-common-protos                 1.74.0
h11                                      0.16.0
httpcore                                 1.0.9
httpx                                    0.28.1
httpx-sse                                0.4.3
idna                                     3.11
importlib_metadata                       8.7.1
ipykernel                                7.2.0
ipython                                  9.11.0
ipython_pygments_lexers                  1.1.1
jaraco.classes                           3.4.0
jaraco.context                           6.1.0
jaraco.functools                         4.4.0
jedi                                     0.19.2
Jinja2                                   3.1.6
jiter                                    0.13.0
jsonpatch                                1.33
jsonpointer                              3.0.0
jsonref                                  1.1.0
jsonschema                               4.26.0
jsonschema-path                          0.3.4
jsonschema-specifications                2025.9.1
jupyter_client                           8.8.0
jupyter_core                             5.9.1
keyring                                  25.7.0
langchain                                1.2.10
langchain-core                           1.2.27
langchain-openai                         1.1.12
langfuse                                 4.3.1
langgraph                                1.0.10
langgraph-checkpoint                     4.0.1
langgraph-prebuilt                       1.0.8
langgraph-sdk                            0.3.9
langsmith                                0.7.9
llvmlite                                 0.46.0
lupa                                     2.6
markdown-it-py                           4.0.0
MarkupSafe                               3.0.3
matplotlib-inline                        0.2.1
mcp                                      1.26.0
mdurl                                    0.1.2
more-itertools                           10.8.0
mpmath                                   1.3.0
nest-asyncio                             1.6.0
networkx                                 3.6.1
numba                                    0.64.0
numpy                                    2.4.2
openai                                   2.30.0
openai-whisper                           20250625
openapi-pydantic                         0.5.1
opentelemetry-api                        1.41.0
opentelemetry-exporter-otlp-proto-common 1.41.0
opentelemetry-exporter-otlp-proto-http   1.41.0
opentelemetry-exporter-prometheus        0.60b1
opentelemetry-instrumentation            0.60b1
opentelemetry-proto                      1.41.0
opentelemetry-sdk                        1.41.0
opentelemetry-semantic-conventions       0.62b0
orjson                                   3.11.7
ormsgpack                                1.12.2
packaging                                26.0
parso                                    0.8.6
pathable                                 0.4.4
pathvalidate                             3.3.1
pip                                      26.0.1
platformdirs                             4.5.1
prometheus_client                        0.24.1
prompt_toolkit                           3.0.52
protobuf                                 6.33.6
psutil                                   7.2.2
pure_eval                                0.2.3
py-key-value-aio                         0.3.0
py-key-value-shared                      0.3.0
pycparser                                3.0
pydantic                                 2.12.5
pydantic_core                            2.41.5
pydantic-settings                        2.8.1
pydocket                                 0.16.6
Pygments                                 2.19.2
PyJWT                                    2.10.1
pyperclip                                1.11.0
python-dateutil                          2.9.0.post0
python-dotenv                            1.2.1
python-json-logger                       4.0.0
python-multipart                         0.0.22
pywin32                                  311
pywin32-ctypes                           0.2.3
PyYAML                                   6.0.3
pyzmq                                    27.1.0
redis                                    7.1.0
referencing                              0.36.2
regex                                    2026.2.28
requests                                 2.32.5
requests-toolbelt                        1.0.0
rich                                     14.3.1
rich-rst                                 1.3.2
rpds-py                                  0.30.0
setuptools                               82.0.0
shellingham                              1.5.4
six                                      1.17.0
sniffio                                  1.3.1
sortedcontainers                         2.4.0
sse-starlette                            3.2.0
stack-data                               0.6.3
starlette                                0.52.1
sympy                                    1.14.0
tenacity                                 9.1.4
tiktoken                                 0.12.0
torch                                    2.10.0
tornado                                  6.5.5
tqdm                                     4.67.3
traitlets                                5.14.3
typer                                    0.21.1
typing_extensions                        4.15.0
typing-inspection                        0.4.2
urllib3                                  2.6.3
uuid_utils                               0.14.1
uvicorn                                  0.40.0
wcwidth                                  0.6.0
websockets                               16.0
wrapt                                    1.17.3
xxhash                                   3.6.0
zipp                                     3.23.0
zstandard                                0.25.0
